Allow diag.t to pass tests for multi-line warnings.
It's been sending the full warning text to t/test.pl's pass()
function.  If there are newlines in the the warning, t/TEST (though
not t/harness) will spew "Unexpected output at ..." messages and
fail the test.  In this context, the warning is just a message
printed by the test infrastructure, so we might as well use the
version of the warning that's already had newlines replaced with
spaces.