Unicode::UCD: Use 6.1 format for Name_Alias
The Name_Alias property is changing significantly in Unicode 6.1.  It
will have an extra value for each alias (and there will be a lot more
aliases).  That extra value is considered informative, and gives the
type of alias this is.  prop_invmap() should return both the alias and
its type.  This commit changes it so that even in 6.0, it will return
what it is going to return in 6.1 (for the 6.0 aliases).

This function has not been released in a stable Perl version.  Thus,
this gets the ultimate format into Perl before anyone has come to depend
on the earlier one.

squish