Unicode::UCD: Allow prop_invmap() to work on non-compact binary properties
It may be that the tables input to prop_invmap() are not in the most
compact form.  There was a problem coping with these on binary
properties; which this commit fixes by realizing that a binary property
omits its 'Y' property value.

The .t does not pass these tests when the input is not compact, though.