Unicode::UCD: Special case prop_aliases() 'indic'
This is in preparation for Unicode 6.1 which adds a property whose name
begins with 'indic', as in Indic languages.  This creates a conflict
with Perl's longstanding use of 'in' as a prefix in a property name.
"\p{infoo}" is a synonym for "\p{blk=foo}".  The conflict is resolved
here with a special case.