[perl #104288] Die with ‘Unrecognized switch’ on #! line
If a switch such as -x that cannot occur on the shebang line is used
there, perl dies with ‘Can’t emulate...’.

If an unrecognised switch is used on the command line, perl dies with
‘Unrecognized switch...’.

It just happens to use the former error message for unrecognized
switches on the shebang line, which can be confusing:

$ perl -e '#!perl -G'
Can't emulate -G on #! line at -e line 1.

This commit changes it to output the same message as command-line
switches for those that are unrecognised wherever they are used.