strlcpy fix-ups in vms/vms.c following a35dcc95dd24.
In three places I was using the intended string length rather than
the buffer length as the length argument, and in one place I was
assuming the return value was the number of non-null bytes copied,
which is not true when you are intentionally copying only part of
the source string.  If strlcpy is supposed to be idiot-proof, they
clearly didn't anticipate the superior form of idiot I've proven
myself to be.