Update method caches for non-void stash elem deletions
Commit 71be2cb (inseparable changes from patch from perl5.003_13 to
perl5.003_14) added code to hv_free_ent to reset method caches if
necessary.

Later, a bug fix in the deletion code (f50383f5) made it necessary for
the value in the HE to be set to &PL_sv_placeholder, so it wouldn’t
free the sv just yet.  So the method cache code (which by then had
changed from using PL_sub_generation to using mro_method_changed_in)
got repeated in S_hv_delete_common.

The only problem with all this is that hv_free_ent was the wrong place
to put it to begin with.  If delete is called in non-void context,
the sv in it is not freed just yet, but returned; so hv_free_ent was
already being called with a HE pointing to &PL_sv_placeholder.

Commit f50383f5 only added the mro code for the void case, to avoid
changing existing behaviour when rearranging the code.

It turns out it needs to be done in S_hv_delete_common uncon-
ditionally.

Changing a test in t/op/method.t to use ()=delete instead of delete
makes it fail, because method caches end up going stale.  See the test
in the diff.