Remove method cache reset from hv.c:hv_free_ent
It hasn’t been necessary since commit f50383f5.  Before that it wasn’t
sufficient.  See commit 5743f2a for details.

If a hash element is being deleted, S_hv_delete_common takes care of
this.  If a hash is being freed or cleared, hv_undef or hv_clear takes
care of it.