Avoid t/re/regexp.t failing on miniperl when displaying TODO test output.
Change 1e9285c2ad54ae39 refactored Data::Dumper to load on miniperl.
t/re/regexp.t attempts to load Data::Dumper (in an eval) to display failure
output, including the failure of TODO tests. Hence Data::Dumper is now loaded
without error as part of minitest, so regexp.t then attempts to use
Data::Dumper to output better diagnostics. This fails (hard) because
Data::Dumper attempts to load Scalar::Util, which attempts to load B, which
bails out because this is miniperl.

It's not obvious that there's a 100% solution here that gets full-on
Data::Dumper functionality for miniperl.