Simplify gv:S_maybe_add_coresub
It was working around the fact that newATTRSUB expects to be able to
look up the GV by name.

And for speed, it was going through hoops to try to avoid creating
extra SVs holding the name unnecessarily.

By tweaking newATTRSUB to accept a GV instead of a name, we can sim-
plify not only S_maybe_add_coresub but all its callers, too.