[perl #107366] Don’t clone GVs during thread join
unless they are orphaned.

This commit stops globs that still reside in their stashes from being
cloned during a join.

That way, a sub like sub{$::x++}, when cloned into a subthread and
returned from it, will still point to the same $::x.

This commit takes the conservative approach of copying on those globs
that can be found under their names in the original thread.

While this doesn’t work for all cases, it’s probably not possible to
make it work all the time.