utf8_heavy.pl: Skip unnecessary work for official properties
The tables that mktables generates are well behaved, and so the checks
and sorting that are done for user-defined properties may be skipped.

tainting needs to be preserved because $list can be passed in already
tainted.

This is also in preparation for Unicode 6.1, in which one table will
legitimately have duplicate entries that the old code removed.