[perl #95548] Returned magical temps are not copied
return and leavesub, for speed, were not copying temp variables with a
refcount of 1, which is fine as long as the fact that it was not cop-
ied is not observable.

With magical variables, that *can* be observed, so we have to forego
the optimisation and copy the variable if it’s magical.

This obviously applies only to rvalue subs.