[perl #90030] sort with no arguments
This eliminates the error ‘sort is now a reserved word’, which was
added for the sake of code that does close(sort) at about the time
that sort became a keyword.

As Tom Christiansen pointed out, it has been long enough.

This error only occurred with ) or ; after the keyword.  In other
cases with no arguments, like {sort} and (sort,0), it was treated as
an empty list.

So this commit makes it follow that.

Since the tests triggered it, this commit also fixes a crash caused by
commit 540dd770, which, when seeing whether it can optimise something
like ‘@a = sort @a’, assumes that there is something after the sort,
and crashes when there isn’t, as in {@a = sort}.