[perl #92706] In PerlIO::Scalar::seek, don’t assume SvPOKp
Otherwise we get assertion failures.

In fact, since seeking might be just for reading, we can’t coerce and
SvGROW either.

In fact, since the scalar might be modified between seek and write,
there is no *point* in SvGROW during seek, even for SvPOK scalars.

PerlIO::scalar assumes in too many places that the scalar it is using
is its own private scalar that nothing else can modify.  Nothing could
be farther from the truth.

This commit moves the zero-fill that usually happens when seeking past
the end from seek to write.  During a write, if the current position
is past the end of the string, the intervening bytes are zero-filled
at that point, since the seek hasn’t done it.