[perl #107440] Save av/hv on mortals stack when clearing
In pp_undef and pp_aassign, we should put the av or hv that is being
cleared on the mortals stack (with an increased refcount), so that
destructors fired during the clearing do not free the av or hv.

I was going to put this in av_undef, etc., but pp_aassign also needs
to access the aggregate after clearing it.  We still get a crash with
that approach.

Putting the aggregate on the mortals stack in av_undef, av_clear and
h_freeentries would work, too, but might cause the aggregate to leak
too far.  That may cause problems, e.g., if it is %^H, because it may
last until the end of the current compilation unit.

Directly inside a runloop (in a pp function), it should be OK to use
the mortals stack, as it *will* be cleared ‘soon’.  This seems the
least intrusive approach.