[perl #67490] Don’t call DELETE on scalar-tied elem
This little snippet:

sub TIESCALAR{bless[]}
sub STORE{}
tie $^H{foo}, '';
$^H{foo} = 1;
delete $^H{foo};

dies with ‘Can't locate object method "DELETE"...’.

This bug was introduced for %^H by commit b3ca2e834c, but it is actu-
ally an older bug that already affected %ENV before that.

Clear-magic on a scalar is only called when it is an element of a mag-
ical aggregate.

For hashes, this clear-magic is called whenever the hash itself
is RMAGICAL.

Tied scalars and elements of tied aggregates use the same magic vta-
ble, under the assumption that mg_clear will never be called on a tied
scalar.  That assumption is wrong.

Commit b3ca2e834c is the one that made %^H magical, which is why it
caused this problem for %^H.

The obvious solution, giving tied scalars their own vtable, is not as
simple as it sounds, because then tied scalars are no longer RMAGICAL,
and at least some of the tie code assumes that they are.

So the easiest fix is to skip the DELETE call in Perl_magic_clearpack
if the type of magic is PERL_MAGIC_tiedscalar.