[perl #44895] += warning on uninit magic var
The only uses of USE_LEFT in core now occur when SvGETMAGIC has
already been called.  So returning true for magical SVs is not neces-
sary.  In fact, it was never correct.

Also, the code in do_vop (which handles bitwise operations on strings)
to avoid an uninitialized warning had the same buggy SvGMAGICAL check.

Now, the warning from $uninit += 1 is suppressed for all undefined
vars, not just amagical ones.

This causes 6 to-do tests in assignwarn.t to pass.