[perl #40333] Stop overload::Overloaded from calling ->can
It’s possible, and too easy, for classes to define a can method to
deal with AUTOLOAD, without taking overloading into account.  Since
AUTOLOAD is the main reason for overriding can, and since overloading
does not respect autoloading, can overrides should not be expected to
deal with it.

Since overload.pm already has a mycan function that fits this purpose,
this commit changes Overloaded to use that.

The test includes an example of a class structure that the previous
Overloaded implementation could not handle.