[perl #24237] @& should not stop $& from working
Mentioning $& in a program slows everything down, because it force
regular expressions to do a pre-match copy.

It used to happen for any symbol named &, e.g., @& and %&.  This was
changed in commit b4a9608f339, but that commit did not take into
account that the code path in question is only followed on creation of
the *& glob.

It should still be applying magic to $&, even if it is not setting
PL_sawampersand.  The other place in gv_fetchpvn_flags that magical-
ises scalars (which currently handles %- %+ %! $] and @ISA), should
also turn on PL_sawampersand for $&.

All of the above applies to $' and $` as well.