In pp_sys.c:pp_fttext, don’t call cGVOP_gv on an UNOP
Otherwise we might get a crash.  cGVOP_gv is only valid when the
OPf_REF flag is set.  In either case, gv already holds the GV we
want anyway.

This code has been buggy this way since this commit:

commit 5f05dabc4054964aa3b10f44f8468547f051cdf8
Author: Perl 5 Porters <perl5-porters@africa.nicoh.com>
Date:   Thu Dec 19 16:44:00 1996 +1200

    [inseparable changes from patch from perl5.003_11 to perl5.003_12]

but apparently has not actually caused a crash until just now.

I was trying to add a test for another bug (fixed in the next commit),
and it happened to trigger this one.  My attempt to reduce this to
something small and reproducible failed.