Don't set PL_perl_destruct_level = 2 as a side effect of use threads;
There's no reason to enable full global destruction in the *top most thread*,
which is what the code used to do by setting PL_perl_destruct_level in a
BOOT block. The top most thread of an interpreter that has created threads
is no different to the only thread of an interpreter that has never created
threads, and we don't force full destruction there.

Instead, it only needs to be enabled in child threads, to ensure that they
are properly cleaned up.

I'm not confident that there are zero users out there on existing perls,
possibly in embedding scenarios, who are relying on some side effect of
setting destruction globally, so for now only enable this change on newer
perls. (As we expect people to have tested things more carefully before
upgrading their entire perl installation, and can't legislate against all
foolishness if they don't.)