defined *{"!"} should not stop %! from working
This is something I broke when merging is_magical_gv with
gv_fetchpvn_flags.

gv_fetchpvn_flags must make sure its *! glob is present in the sym-
bol table it loads Errno.  If it adds it afterwards it will clobber
another *! that Errno has autovivified and tied in the mean time.