defined *{"+"} should not stop %+ from working
The same applies to %-.

This is something I broke when merging is_magical_gv with
gv_fetchpvn_flags.

gv_fetchpvn_flags must make sure its *+ glob is present in the symbol
table when it loads Tie::Hash::NamedCapture.  If it adds it afterwards
it will clobber another *+ that Tie::Hash::NamedCapture has autovivi-
fied and tied in the mean time.