locale.t: Don't use hard-coded test numbers
This was rather painful to convert the hard-coded numbers into
calculated ones so that tests could be added and subtracted.  The debug
statements were moved to after the last test they described so the test
numbers would be calculated, and a new hash created to deal with
skipping tests and not knowing how many are skipped; otherwise the
current test number is kept track of and incremented as needed.