[perl #104084] Make system call FETCH in parent proc
system() calls FETCH on tied variables in the child process, after
forking, so the FETCH routine sees an intermediate state that is sup-
posed to be an implementation detail on Unix.

The solution is to call get-magic up front on all the arguments, as we
already do under taint mode, and skip get-magic later.

This also means that taint mode no longer causes get-magic to be
called twice for each argument.