[perl #106726] Don’t crash on length(@arr) warning
The RT ticket blames this on 676a678ac, but it was actually commit
579333ee9e3.  676a678ac extended this problem to evals (and modules),
but it already occurred in the main program.

This crashes:

    ./miniperl -Ilib -we 'sub {length my @forecasts}'

because it is trying to find the variable name for the warning in the
CV returned by find_runcv, but this is a *compile-time* warning, so
using find_runcv is just wrong.

It ends up looking for the array in PL_main_cv’s pad, instead of
PL_compcv.