Don’t crash on @a =~ // warning
This is similar to bug #106726, caused by 579333ee9e3 and fixed
by c6fb3f6.

This bug was caused by c6771ab63d, which had exactly the same mistake.
Attempting to use find_runcv at compile time to find a variable name
is not going to work in general.  It only works by accident some
of the time.