Safe.pm: Make sure SWASHNEW is properly loaded
This module was depending on testing code points in the upper Latin1
range causing utf8_heavy.pl.  However a recent performance improvement
caused those code points to skip the loading.  This just changes the
code points to two higher values that cause it to load, and until and if
it changes again, will fix things.