Restore triple-dot directory spec tests in vms/ext/filespec.t.
1fe570cc5e24eecfb07059e53e95fa864bb44142 declared directory
components containing '...' as either 'not translatable' or created
the expectation that each dot should be individually escaped when
translating between Unix and VMS directory specs.  That doesn't
really make sense since in both formats it means any number of
intervening directories, plus there was already code of long
standing that handles it.  So get the tests in this regard back
in line with reality.