mktables: Refactor populating simple case folding tables
These three tables are handled alike; this creates a loop to execute the
same instructions on each of them.  Currently there is so little to do,
that it wouldn't be worth it, except that future commits will add
complications, and this makes those easier to handle.

There is now a test that the input data is sane, and instead of
overwriting a value in a table with a known identical value, we skip
that.  This doesn't save much effort, because most of the work is
looking up the value (which we can now check sanity for), but again will
be useful for future commits.