mktables: Add duplicate tables
This is for backwards compatibility.  Future commits will change these
tables that are generated by mktables to be more efficient.  But the
existence of them was advertised in v5.12 and v5.14, as something a Perl
program could use because the Perl core did not provide access to their
contents.  We can't change the format of those without some notice.

The solution adopted is to have two versions of the tables, one kept in
the original file name has the original format; and the other is free to
change formats at will.

This commit just creates copies of the original, with the same format.
Later commits will change the format to be more efficient.

We state in v5.16 that using these files is now deprecated, as the
information is now available through Unicode::UCD in a stable API.  But
we don't test for whether someone is opening and reading these files; so
the deprecation cycle should be somewhat long;  they will be unused, and
the only drawbacks to having them are some extra disk space and the time
spent in having to generate them at Perl build time.

This commit also changes the Perl core to use the original tables, so
that the new format can be gradually developed in a series of patches
without having to cut over the whole thing at once.