mktables: Include simple mappings in full tables
This changes the case change mapping tables to include the simple
mappings.  This was done in 5.14 for the case folding table.  The full
mappings are contained, as before, in a hash.  Now the simple mappings
they override (when doing multi-char case changing) are added to the
main body of the table, to the already existing simple mappings that
aren't overridden.

If the caller wants to do full mapping, it should look first in the
hash, and only if not found, look in the main body.  If the caller wants
only simple mapping, it ignores the hash.

This is already how the code in utf8.c that reads these tables is
constructed.

The .t is modified to take into account that these code points are now
in the main table body.