Handle directories with dots in fileify under EFS.
When Extended Filename Syntax (EFS) is in effect, it's legal for
directory names to contain dots that are not directory delimiters.
This commit escapes such dots when they are part of the file
portion of the fileified directory spec.

This gets two more of the vms/ext/filespec.t tests passing when
EFS is in effect.